New changes

+++ cross_shield_backend/database/migrations/013_create_add_medications_to_medication_requests_table.ts
@@ -4,12 +4,20 @@ export default class extends BaseSchema {
   protected tableName = 'medication_requests'
 
   public async up() {
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'medications')
+    if (hasColumn) {
+      return
+    }
     this.schema.alterTable(this.tableName, (table) => {
       table.text('medications').nullable()
     })
   }
 
   public async down() {
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'medications')
+    if (!hasColumn) {
+      return
+    }
     this.schema.alterTable(this.tableName, (table) => {
       table.dropColumn('medications')
     })
