+++ cross_shield_backend/database/migrations/014_create_remove_medication_name_from_requests_table.ts
@@ -2,12 +2,20 @@ import { BaseSchema } from '@adonisjs/lucid/schema'
 
 export default class extends BaseSchema {
   async up() {
+    const hasColumn = await this.schema.hasColumn('medication_requests', 'medication_name')
+    if (!hasColumn) {
+      return
+    }
     this.schema.alterTable('medication_requests', (table) => {
       table.dropColumn('medication_name')
     })
   }
 
   async down() {
+    const hasColumn = await this.schema.hasColumn('medication_requests', 'medication_name')
+    if (hasColumn) {
+      return
+    }
     this.schema.alterTable('medication_requests', (table) => {
       table.string('medication_name')
     })
