+++ cross_shield_backend/database/migrations/015_create_remove_quantity_from_requests_table.ts
@@ -2,12 +2,20 @@ import { BaseSchema } from '@adonisjs/lucid/schema'
 
 export default class extends BaseSchema {
   async up() {
+    const hasColumn = await this.schema.hasColumn('medication_requests', 'quantity')
+    if (!hasColumn) {
+      return
+    }
     this.schema.alterTable('medication_requests', (table) => {
       table.dropColumn('quantity')
     })
   }
 
   async down() {
+    const hasColumn = await this.schema.hasColumn('medication_requests', 'quantity')
+    if (hasColumn) {
+      return
+    }
     this.schema.alterTable('medication_requests', (table) => {
       table.integer('quantity').notNullable()
     })
