+++ cross_shield_backend/database/migrations/017_create_add_payload_to_requests_table.ts
@@ -4,12 +4,20 @@ export default class extends BaseSchema {
   protected tableName = 'requests'
 
   public async up() {
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'payload')
+    if (hasColumn) {
+      return
+    }
     this.schema.alterTable(this.tableName, (table) => {
       table.json('payload').nullable()
     })
   }
 
   public async down() {
+    const hasColumn = await this.schema.hasColumn(this.tableName, 'payload')
+    if (!hasColumn) {
+      return
+    }
     this.schema.alterTable(this.tableName, (table) => {
       table.dropColumn('payload')
     })
